Added Standard Idle Animation
I'm using the player prefab to test it which creates a small problem with placing the sprite. The bounds of all standard spites are 96x96 and the way I have it they arent centered like the player sprites are. I'll play with this to find a solution but so far it looks pretty good in the environments we have right now.