Doors Animated and Legacy Assets in
Jason Le's doors (except for the elevator) are in and animated! Their animation states are in the prop animation folder. They have also been prefab'd. Feel free to mess with the prefab, engineers!
Majority of Legacy Assets from the previous concept build should be in and animated now. Will be using this to set dress soon.