Added Interactable Doors.
To use, add 1 BlackoutController to the scene (As long as there are doors, the blackout is necessary under the Canvas object. Then, add the WorkingDoorSet prefab into the scene and place the 2 doors wherever you want. Interacting with them will fade the screen to black and play door animations/sounds, teleport the player to the other door, and fade in and give control back to the player.

For the blackout controller: Added additional functions for fade out and fade in without any other effects. When using this, you do not need a GatherInput variable asked for on the script.